(Menu) Use float for icon coordinates, makes animations look smoother

+++ menu/drivers_display/menu_display_d3d.cpp
@@ -102,7 +102,7 @@ static void menu_display_d3d_blend_end(void)
 }
 
 static void menu_display_d3d_draw(
-      unsigned x, unsigned y,
+      float x, float y,
       unsigned width, unsigned height,
       struct gfx_coords *coords,
       void *matrix_data,
